feat: enhance core functionality with new MotiaPluginContext and API endpoints integration
- Added MotiaPluginContext to streamline plugin development.
- Introduced new EndpointsStream for managing API endpoints.
- Removed deprecated apiEndpoints module to simplify the codebase.
- Updated various plugins to utilize the new context and endpoint management features.
