+++ packages/core/index.ts
@@ -12,7 +12,9 @@ export { StateAdapter } from './src/state/state-adapter'
 export { createMermaidGenerator } from './src/mermaid-generator'
 export { StreamConfig, MotiaStream } from './src/types-stream'
 export { getProjectIdentifier, getUserIdentifier, isAnalyticsEnabled, trackEvent } from './src/analytics/utils'
-export { Motia } from './src/motia'
+export { Motia, MotiaPluginContext } from './src/motia'
 export { NoPrinter, Printer } from './src/printer'
 export { NoTracer } from './src/observability/no-tracer'
 export { config } from './src/config'
+export * from './src/types/schema.types'
+export { StreamAdapter } from './src/streams/adapters/stream-adapter'
