+++ packages/core/src/motia.ts
@@ -1,10 +1,10 @@
-import { Printer } from './printer'
-import { TracerFactory } from './observability'
-import { EventManager, InternalStateManager } from './types'
+import type { Express } from 'express'
 import { LockedData } from './locked-data'
 import { LoggerFactory } from './logger-factory'
-import type { Express } from 'express'
+import { TracerFactory } from './observability'
+import { Printer } from './printer'
 import { StateAdapter } from './state/state-adapter'
+import { EventManager, InternalStateManager } from './types'
 
 export type Motia = {
   loggerFactory: LoggerFactory
@@ -16,4 +16,10 @@ export type Motia = {
 
   app: Express
   stateAdapter: StateAdapter
+}
+
+export type MotiaPluginContext = {
+  app: Express
+  state: StateAdapter
+  lockedData: LockedData
 }
